requirements.txt
pytest>=9.0.2
selenium>=4.40.0
undetected-chromedriver>=3.5.5
requests>=2.32.5
python-dotenv>=1.2.1
allure-pytest>=2.15.3